added alternative labelling method (old one proved to be better), rolling std function and calculated labels from original px in training
